refactor: :recycle: Create safe bet handling, add startup modules
Added method in `MatchupManager` to be used to verify existing bets on a specified matchup. This is to be used in conjunction with `gameHeartbeat` - as the latter module will process closing game channels, bets, and removing them from the score db.
The score db is now used to determine if a game is closed or not to reduce API calls to The-Odds API, while hopefully maintaining a short amount of time to process bets after a game has ended.
By avoiding deleting open bets before games are officially closed, thie app will repeatedly check the same game along with the bets for it until there are none.
In addition, the db responsible for letting us know if X user has a pending bet will clear on startup.
Cron Ranges have been removed in favor of using the initially mentioned system of score db => finds finished game => processes bets.
Added Odds collection to scheduled modules.